backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

sqlalchemy==2.0.36
psycopg[binary]==3.2.3
python-dotenv==1.0.1
